Updating MAGIC DB struct
Change of recast skill cooldown timer from tinyint to smallint. Because tinyint supports 255 as a max value and there are some skills that have 600 as cooldown timer. Example: Sacrifice with magicID
106660 (Sacrifice your own self to fill the HP one party member)

+++ src/server/Ebenezer/MagicTableSet.h
@@ -30,7 +30,7 @@ class CMagicTableSet : public CRecordset {
     BYTE    m_ItemGroup;
     long    m_UseItem;
     BYTE    m_CastTime;
-    BYTE    m_ReCastTime;
+    int    m_ReCastTime;
     BYTE    m_SuccessRate;
     BYTE    m_Type1;
     BYTE    m_Type2;
